+++ build-tools/packages/build-tools/src/repoPolicyCheck/handlers/npmPackages.ts
@@ -742,4 +742,88 @@ export const handlers: Handler[] = [
 				: undefined;
 		},
 	},
+	{
+		name: "npm-package-json-test-scripts",
+		match,
+		handler: (file, root) => {
+			let json;
+
+			try {
+				json = JSON.parse(readFile(file));
+			} catch (err) {
+				return "Error parsing JSON file: " + file;
+			}
+
+			const packageDir = path.dirname(file);
+			const scripts = json.scripts;
+			if (
+				scripts !== undefined &&
+				Object.keys(scripts).some((name) => name.startsWith("test"))
+			) {
+				// Found test script
+				return undefined;
+			}
+
+			const testDir = path.join(packageDir, "src", "test");
+			if (fs.existsSync(testDir)) {
+				const info = fs.readdirSync(testDir, { withFileTypes: true });
+				if (
+					info.some(
+						(e) =>
+							path.extname(e.name) === ".ts" ||
+							(e.isDirectory() && e.name !== "types"),
+					)
+				) {
+					return "Test files exists but no test scripts";
+				}
+			}
+
+			const dep = ["mocha", "@types/mocha", "jest", "@types/jest"];
+			if (
+				(json.dependencies &&
+					Object.keys(json.dependencies).some((name) => dep.includes(name))) ||
+				(json.devDependencies &&
+					Object.keys(json.devDependencies).some((name) => dep.includes(name)))
+			) {
+				return `Package has one of "${dep.join()}" dependency but no test scripts`;
+			}
+		},
+	},
+	{
+		name: "npm-package-json-script-mocha-config",
+		match,
+		handler: (file, root) => {
+			let json;
+
+			try {
+				json = JSON.parse(readFile(file));
+			} catch (err) {
+				return "Error parsing JSON file: " + file;
+			}
+
+			const scripts = json.scripts;
+			if (scripts === undefined) {
+				return undefined;
+			}
+			const mochaScriptName = scripts["test:mocha"] !== undefined ? "test:mocha" : "test";
+			const mochaScript = scripts[mochaScriptName];
+
+			if (!mochaScript || !mochaScript.startsWith("mocha")) {
+				// skip irregular test script for now
+				return undefined;
+			}
+
+			const packageDir = path.dirname(file);
+			const mochaRcNames = [".mocharc", ".mocharc.js", ".mocharc.json", ".mocharc.cjs"];
+			const mochaRcName = mochaRcNames.find((name) =>
+				fs.existsSync(path.join(packageDir, name)),
+			);
+
+			if (mochaRcName === undefined) {
+				if (!mochaScript.includes(" --config ")) {
+					return "Missing config arguments";
+				}
+			}
+		},
+	},
 ];