+++ packages/loader/driver-utils/src/test/parallelRequests.spec.ts
@@ -5,7 +5,7 @@
 
 import { strict as assert } from "assert";
 import { MockLogger } from "@fluidframework/telemetry-utils";
-import { unreachableCase } from "@fluidframework/common-utils";
+import { unreachableCase } from "@fluidframework/core-utils";
 import { ParallelRequests } from "../parallelRequests";
 
 enum HowMany {