+++ packages/runtime/container-runtime/src/summary/runningSummarizer.ts
@@ -11,14 +11,9 @@ import {
 	createChildLogger,
 	UsageError,
 } from "@fluidframework/telemetry-utils";
-import {
-	assert,
-	delay,
-	Deferred,
-	PromiseTimer,
-	TypedEventEmitter,
-} from "@fluidframework/common-utils";
-import { DriverErrorType } from "@fluidframework/driver-definitions";
+import { assert, delay, Deferred, PromiseTimer } from "@fluidframework/core-utils";
+import { TypedEventEmitter } from "@fluid-internal/client-utils";
+import { DriverErrorTypes } from "@fluidframework/driver-definitions";
 import { ISequencedDocumentMessage, MessageType } from "@fluidframework/protocol-definitions";
 import { ISummaryConfiguration } from "../containerRuntime";
 import { opSize } from "../opProperties";
@@ -316,7 +311,7 @@ export class RunningSummarizer extends TypedEventEmitter<ISummarizerEvents> impl
 						// by calling `refreshLatestSummaryAckFromServer` and we will be fine.
 						const isIgnoredError =
 							isFluidError(error) &&
-							error.errorType === DriverErrorType.fileNotFoundOrAccessDeniedError;
+							error.errorType === DriverErrorTypes.fileNotFoundOrAccessDeniedError;
 
 						summaryLogger.sendTelemetryEvent(
 							{
@@ -606,7 +601,7 @@ export class RunningSummarizer extends TypedEventEmitter<ISummarizerEvents> impl
 				this.beforeSummaryAction();
 			},
 			async () => {
-				return this.mc.config.getBoolean("Fluid.Summarizer.TryDynamicRetries")
+				return this.mc.config.getBoolean("Fluid.Summarizer.UseDynamicRetries")
 					? this.trySummarizeWithRetries(reason)
 					: this.trySummarizeWithStaticAttempts(reason);
 			},