+++ packages/runtime/container-runtime/src/test/gc/garbageCollection.spec.ts
@@ -26,7 +26,7 @@ import {
 	tagCodeArtifacts,
 	createChildLogger,
 } from "@fluidframework/telemetry-utils";
-import { Timer } from "@fluidframework/common-utils";
+import { Timer } from "@fluidframework/core-utils";
 import {
 	concatGarbageCollectionStates,
 	GarbageCollector,
@@ -53,10 +53,9 @@ import {
 	dataStoreAttributesBlobName,
 	IContainerRuntimeMetadata,
 	metadataBlobName,
-	RefreshSummaryResult,
 } from "../../summary";
 import { pkgVersion } from "../../packageVersion";
-import { configProvider, parseNothing } from "./gcUnitTestHelpers";
+import { configProvider } from "./gcUnitTestHelpers";
 
 type GcWithPrivates = IGarbageCollector & {
 	readonly configs: IGarbageCollectorConfigs;
@@ -931,86 +930,6 @@ describe("Garbage Collection Tests", () => {
 					"Expecting 1 deleted node",
 				);
 			});
-
-			it("reads all GC data from when refreshing from snapshot with same GC version", async () => {
-				const garbageCollector = createGCOverride(stableGCVersion);
-				await garbageCollector.initializeBaseState();
-
-				// Get a snapshot with the current GC version and refresh latest summary state from it.
-				const { snapshotTree, gcBlobsMap } = getSnapshotWithGCVersion(stableGCVersion);
-				const refreshSummaryResult: RefreshSummaryResult = {
-					latestSummaryUpdated: true,
-					wasSummaryTracked: false, // Indicates that state has to be updated from the snapshot in the result.
-					summaryRefSeq: 0,
-					snapshotTree,
-				};
-				await garbageCollector.refreshLatestSummary(
-					undefined,
-					refreshSummaryResult,
-					async <T>(id: string) => gcBlobsMap.get(id) as T,
-				);
-
-				// The latest summary state should all be updated from the snapshot.
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData !== undefined,
-					"Latest summary data not updated",
-				);
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData.serializedGCState !==
-						undefined,
-					"Latest summary GC state not updated",
-				);
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData.serializedTombstones !==
-						undefined,
-					"Latest summary tombstone state not updated",
-				);
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData
-						.serializedDeletedNodes !== undefined,
-					"Latest summary deleted nodes not updated",
-				);
-			});
-
-			it("discards all GC data from when refreshing from snapshot with different GC version", async () => {
-				const garbageCollector = createGCOverride(stableGCVersion);
-				await garbageCollector.initializeBaseState();
-
-				// Get a snapshot with different GC version from current and refresh latest summary state from it.
-				const { snapshotTree, gcBlobsMap } = getSnapshotWithGCVersion(stableGCVersion + 1);
-				const refreshSummaryResult: RefreshSummaryResult = {
-					latestSummaryUpdated: true,
-					wasSummaryTracked: false, // Indicates that state has to be updated from the snapshot in the result.
-					summaryRefSeq: 0,
-					snapshotTree,
-				};
-				await garbageCollector.refreshLatestSummary(
-					undefined,
-					refreshSummaryResult,
-					async <T>(id: string) => gcBlobsMap.get(id) as T,
-				);
-
-				// Only the deleted nodes state should be updated from this snapshot since the GC version changed.
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData !== undefined,
-					"Latest summary data not updated",
-				);
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData.serializedGCState ===
-						undefined,
-					"Latest summary GC state should now be undefined",
-				);
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData.serializedTombstones ===
-						undefined,
-					"Latest summary tombstone state should now be undefined",
-				);
-				assert(
-					garbageCollector.summaryStateTracker.latestSummaryData
-						.serializedDeletedNodes !== undefined,
-					"Latest summary GC deleted nodes should be updated",
-				);
-			});
 		});
 
 		it("generates both inactive and sweep ready events when nodes are used after time out", async () => {
@@ -1178,16 +1097,10 @@ describe("Garbage Collection Tests", () => {
 				"Deleted nodes state should be a handle",
 			);
 
-			const refreshSummaryResult: RefreshSummaryResult = {
-				latestSummaryUpdated: true,
-				wasSummaryTracked: true,
-				summaryRefSeq: 0,
-			};
-			await garbageCollector.refreshLatestSummary(
-				undefined,
-				refreshSummaryResult,
-				parseNothing,
-			);
+			await garbageCollector.refreshLatestSummary({
+				isSummaryTracked: true,
+				isSummaryNewer: true,
+			});
 
 			// Run GC and summarize again. The whole GC summary should now be a summary handle.
 			await garbageCollector.collectGarbage({});
@@ -1781,12 +1694,10 @@ describe("Garbage Collection Tests", () => {
 
 			checkGCSummaryType(tree1, SummaryType.Tree, "first");
 
-			await garbageCollector.refreshLatestSummary(
-				undefined,
-				{ wasSummaryTracked: true, latestSummaryUpdated: true, summaryRefSeq: 0 },
-				parseNothing,
-			);
-
+			await garbageCollector.refreshLatestSummary({
+				isSummaryTracked: true,
+				isSummaryNewer: true,
+			});
 			await garbageCollector.collectGarbage({});
 			const tree2 = garbageCollector.summarize(fullTree, trackState);