+++ packages/utils/telemetry-utils/src/sampledTelemetryHelper.ts
@@ -9,7 +9,7 @@ import {
 	ITelemetryProperties,
 	IDisposable,
 } from "@fluidframework/core-interfaces";
-import { performance } from "@fluidframework/common-utils";
+import { performance } from "@fluid-internal/client-utils";
 import { ITelemetryLoggerExt } from "./telemetryTypes";
 
 /**